+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to the "sentry-issues" extension will be documented in this 
 
 - Add icon for VSCode marketplace
 - Show progress bar on issue view while refreshing
+- Display `Show Output` button when error occurs
 
 ## 1.0.0 `18 July 2023`